codeExtract

Returns aggregated and cleaned GAMS code together with declaration matrix

Usage

codeExtract(codeFiles, name)

Arguments

codeFiles

A vector of file names of GAMS code files.

name

A name indicating what collection of code files this is (e.g. module name)

Value

A list with two elements: code and declarations. Code contains the cleaned up gams code and declarations contains the declarations matrix as returned by readDeclarations

See also

Author

Jan Philipp Dietrich