gapsfilter {GMS}	R Documentation

Filter data series based on gaps

Description

The data is filtered based on a gap detection. If the gap between the sorted data becomes too large the data is suspect. A different setting for the upper and lower half can be used (seperated at the median).

Usage

gapsfilter(location, measurement, sensor, gap)

Value

Returns a list with filtered data.

Author(s)

Marieke Dirksen, marieke.dirksen@knmi.nl

[Package GMS version 0.1 Index]