gms: 'GAMS' Modularization Support Package

A collection of tools to create, use and maintain modularized model code written in the modeling language 'GAMS' (https://www.gams.com/). Out-of-the-box 'GAMS' does not come with support for modularized model code. This package provides the tools necessary to convert a standard 'GAMS' model to a modularized one by introducing a modularized code structure together with a naming convention which emulates local environments. In addition, this package provides tools to monitor the compliance of the model code with modular coding guidelines.

See also

Author

Maintainer: Jan Philipp Dietrich dietrich@pik-potsdam.de (ORCID) (Potsdam Institute for Climate Impact Research)

Authors:

David Klein (Potsdam Institute for Climate Impact Research)

Anastasis Giannousakis

Felicitas Beier (Potsdam Institute for Climate Impact Research)

Johannes Koch (Potsdam Institute for Climate Impact Research)

Lavinia Baumstark (Potsdam Institute for Climate Impact Research)

Mika Pflüger

Oliver Richters (Potsdam Institute for Climate Impact Research)